Move test to /tests/unit directory
