+++ tests/unit/Entities/DatabaseInfoTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Matriphe\Larinfo\Tests\Entities;
+namespace Matriphe\Larinfo\Tests\unit\Entities;
 
 use Illuminate\Database\Capsule\Manager;
 use Illuminate\Database\Connection;
