+++ tests/unit/Entities/IpAddressCheckerCheckerTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Matriphe\Larinfo\Tests\Entities;
+namespace Matriphe\Larinfo\Tests\unit\Entities;
 
 use Matriphe\Larinfo\Entities\IpAddressChecker;
 use PHPUnit\Framework\TestCase;
